Handle non-default branches for metadata repository
This change updates the metadata class to check out the branch configured for the metadata
repository when it clones the repository initially, and then to conduct only fast-forward pulls
from that branch on updates.

Previously, the clone operation would check out the default branch for the metadata repository
(typically "main" or "master"). When it then pulls updates on the configured branch, if that branch
has diverged from the main branch, then the pull command attempts to make a merge commit. In a
containerized runtime environment, the Git user data is not configured, so this commit attempt
fails.