Scope requests to view and edit vote record to only the specified election
The controllers to view and to edit a voter's submission in an election now
search for that vote record within the scope of the ID of the election.

Otherwise, the controller will find the first vote that user submitted
for any election in the system and attempt to decrypt that vote using
the supplied password.